fix file merge & delete old file scheduler

+++ app/utils/file_util.py
@@ -10,19 +10,17 @@ import re
 def get_device_ids():
     data_storage_path =  Path(Config.DATA_STORAGE)
     # 디렉토리 목록을 리스트로 반환
-    device_ids = [str(dir) for dir in data_storage_path.iterdir() if dir.is_dir()]
-    return device_ids
+    device_ids_dir_path = [str(dir) for dir in data_storage_path.iterdir() if dir.is_dir()]
+    return device_ids_dir_path
 
-def merge_files(device_id: str):
+def merge_files(device_id_dir_path: str):
     date_to_files = {}
-    all_columns = []
 
-    device_id = device_id.split("/")[-1]
+    device_id = device_id_dir_path.split("/")[-1]
     folder_path = f"{Config.DATA_STORAGE}/{device_id}"
     all_files = glob.glob(os.path.join(folder_path, f"{device_id}_*.csv"))
 
     date_pattern = re.compile(r"_(\d{8})\d*\.csv$")
-
     for file_path in all_files:
         match = date_pattern.search(file_path)
         if match:
@@ -33,6 +31,7 @@ def merge_files(device_id: str):
 
     for date, files in date_to_files.items():
         files.sort()  
+        all_columns = []
         merged_df = pd.DataFrame()
         for file_path in files:
             try:
@@ -40,14 +39,24 @@ def merge_files(device_id: str):
                 all_columns.append(list(df.columns))
             except pd.errors.EmptyDataError:
                 logging.warning(f"Empty file: {file_path}")
+                continue
 
         merged_filename = f"{folder_path}/{device_id}_{date}.csv"
         if all_columns:
             merged_df = pd.DataFrame(all_columns)
             merged_df.to_csv(merged_filename, index=False, header=False)
             logging.info(f"{len(files)} files merged into {merged_filename}")
-    
-        for file_path in files:
+
+def delete_old_files(device_id_dir_path: str):
+
+    device_id = device_id_dir_path.split("/")[-1]
+    folder_path = f"{Config.DATA_STORAGE}/{device_id}"
+    all_files = glob.glob(os.path.join(folder_path, f"{device_id}_*.csv"))
+
+    date_time_pattern = re.compile(r"_(\d{14})\.csv$")
+
+    for file_path in all_files:
+        if date_time_pattern.search(file_path):
             try:
                 os.remove(file_path)
             except Exception as e:
